requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Testing==0.8.1
greenlet==3.0.3
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
pluggy==1.4.0
PyJWT==2.8.0
pysqlite3==0.5.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
tomli==2.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
